+++ CHANGELOG.md
@@ -23,6 +23,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Renamed salary variables to 'programSalary,' 'schoolSalary,' and 'nationalSalary' to distinguish data sources
 - Adjust the update_ipeds script to create new schools first, so they get updated with IPEDS data too
 - Tweaked load_programs to fall back to windows-1252
+- Added new constant `constantsYear` and changed API and salary year vars to `apiYear` and `salaryYear` 
 
 ## 2.1.4
 - Added string checking for program codes